A lot of modifications to remove the warnings at the compilation. Mainly, it is a series of cast in REAL in the fortran code, as well as promotting variable from REAL to DOUBLE PRECISION. This had the consequence of changing slightly the reference roadcast. This change is the order of magnitude that a bit flit would produce in the model.

+++ src/model/grille.f
@@ -100,8 +100,8 @@
 *     ratio: transition ratio of transfert CS -> C and KS -> Ko
 ***
       INTEGER INTER, NMAX
-      REAL YPG(n), YPT(n), C(n), Ko(n)
-      REAL dd, ratio, CC, DY
+      DOUBLE PRECISION YPG(n), YPT(n), C(n), Ko(n)
+      DOUBLE PRECISION dd, ratio, CC, DY
       DOUBLE PRECISION dpFluxDepth(n)
 
       LOGICAL bSilent
