+++ src/model/initial.f
@@ -115,28 +115,28 @@
       next = 1
       now = 2
       do j=1,iref
-         T(j,now) = ITP(j)
+         T(j,now) = REAL(ITP(j))
       end do
 
       do i = DEB,FIN
          G(0) = 0.0
-         G(1) = dpConductivity(1) * ( T(2,now) - T(1,now) )
-         T(1,next) = TSO(i)
+         G(1) = REAL(dpConductivity(1)) * ( T(2,now) - T(1,now) )
+         T(1,next) = REAL(TSO(i))
          do j=2,iref-1
-            G(j) = dpConductivity(j) * ( T(j+1,now) - T(j,now) )
+            G(j) = REAL(dpConductivity(j)) * ( T(j+1,now) - T(j,now) )
          end do
          do j=2,iref-1
-            T(j,next) = T(j,now)+DT*(dpCapacity(j)*( G(j) - G(j-1)))
+            T(j,next) = T(j,now)+REAL(DT*(dpCapacity(j)*(G(j)- G(j-1))))
          end do
          if ( SWO(i,1) .eq. 1 ) then
-            T(ir40,next) = TUO(i)
+            T(ir40,next) = REAL(TUO(i))
          end if
          if ( FLAT .and. SWO(i,2) .eq. 1 ) then
 *        BC: underside temp. is air temp
-            T(iref,next) = TAO(i)
+            T(iref,next) = REAL(TAO(i))
          else
 *        BC: no flux ( G(iref) = 0.0 )
-            T(iref,next) = T(j,now) - DT*dpCapacity(j)*G(iref-1)
+            T(iref,next) = T(j,now) - REAL(DT*dpCapacity(j)*G(iref-1))
          end if
          next = 3 - next
          now = 3 - now
